(major) Reduce memory usage by up to 50%
Refactor map_solve.mata , removing dead code and splitting the code that
obtains the fixed effects.

Now map_solve copies the variables from Stata into Mata in blocks
respecting poolsize(), instead of together.

This means that we are not keeping two copies of the entire dataset
during that brief moment.