Fix #94 and #95
Issues resolved:

1) qrsolve(XX, XY) suffers from numerical inaccuracies on some cases, so
we fix to qrsolve(X, Y) if we don't have weights

(Note: this might be a bit slower as the XX and XY are already
precomputed. It could also be optimized but for now let's leave it as it
is.

2) The methods used to find out omitted variables were different in
different points. When demeaning, we looked if the regressors were close
to zero, but used absolute values (1e-8) which doesn't work if there is
a huge scaling difference between Y and X. Now we will assumme the var
has been absorbed by the absvars if the ratio z'z/w'w < (1e-9) where
z=old variable and w= demeaned variable. This 1e-9 will also increase
with tolerance() as it's just tolerance*1e-1

We also use invsym() as the main driver for whether to drop or not, as
that is what it's used on the built-in tools. Then the inputs of
qrsolve() will exclude the omitted variables, to prevent any issue.