move weight update to separate method
this allows to keep the same HDFE object but with different weights.
Won't work if we have freq weights and we drop singletons and some
weights change from >0 to 0 or viceversa.

Also wont work if you run one regr without weights and then want to add
weights later (or change from e.g. pw to aw, but this can be done in
other ways)