Try to improve some corner cases when checking convergence
Two main changes:

1. If SSR is very close to zero, it gets edited to zero. Otherwise, on
cases with perfect or near perfect collinearity, we might fail to
converge as we end up in a case where SSR / PreviouSSR = 0 / 0 (or close
to it)
2. Convergence should be invariant to rescaling of the weights. For
instance, if we use pw=weightvar we should get the same result than if
we use pw=1000*weightvar.  However, because convergence is based on the
weighted SSR, this is not the case. This commit fixes the cases with
weights below 1, so setting weights based of pop in thousands will not
give less accurate results than using pop in units. However, we didn't
do the other case (where making all weights larger improves the
accuracy)