Add extra check when computing  dropped regressors
invsym() is not as accurate as qrsolve(), specifically when assigning it a drop ("sweep") order (because otherwise it drops the regressors that minimize roundoff error).

This commit adds an extra check if invsym(xx, 1.k) has determined that we should drop variables. It runs invsym() to check that the same number of variables is being dropped, and uses the latter if there are contradictions.

Cost of the commit: slight slowdown in regressions that drop regressors
Benefits: in some extremely rare cases (such as regressions with incredibly high weights, such as 1e7 or higher), we avoid incorrectly dropping regressors.